+++ electron/services/devTools/index.ts
@@ -1,17 +1,1 @@
-import installExtension, {
-  REACT_DEVELOPER_TOOLS,
-} from 'electron-devtools-installer';
-import { isDev } from '@shared/constants';
-import type { ILogger } from '..';
-
-export function setUpDevtools(logger: ILogger) {
-  return (): void => {
-    if (isDev) {
-      installExtension(REACT_DEVELOPER_TOOLS)
-        .then((name) => {
-          logger.info(`Added Extension: "${name}"`);
-        })
-        .catch(logger.errorWithContext('installing extension'));
-    }
-  };
-}
+export { setUpDevtools } from './devTools';
