+++ electron/services/updater/updater.ts
@@ -1,0 +1,9 @@
+import { autoUpdater } from 'electron-updater';
+import { ILogger } from '../logger';
+
+export function checkForUpdates(logger: ILogger): void {
+  autoUpdater.logger = logger;
+  autoUpdater
+    .checkForUpdatesAndNotify()
+    .catch(logger.errorWithContext('checking for updates'));
+}
