+++ jest.config.js
@@ -6,17 +6,22 @@ const client = './client/**/*.ts?(x)';
 const server = './electron/**/*.ts';
 const shared = './shared/**/*.ts';
 const all = [client, server, shared];
-const escapeHatch = ['!**/github.ts'];
+const escapeHatch = [
+  '!**/index.ts?(x)',
+  '!./electron/main.ts',
+  '!./electron/preload.ts',
+  '!./electron/createWindow.ts',
+];
 module.exports = {
   preset: 'ts-jest',
 
-  collectCoverage: false,
+  collectCoverage: true,
   collectCoverageFrom: all.concat(escapeHatch),
   coverageThreshold: all.reduce(
     (coverage, name) => ({
       ...coverage,
       [name]: {
-        branches: 100,
+        branches: 95,
         functions: 100,
         lines: 100,
         statements: 100,
